Subject: Added logic to avoid duplicate entries in authorized_keys
Problem: Dup keys are added when copying them to remote hosts, this is
normal but sometimes it could flood the authorized_keys with same
information.

Solution: Added logic to avoid duplicate entries in authorized_keys, it could
be that same keys are added If the user runs the script twice ore more; ie.
running in in cron jobs or automation process.

Note: